+++ src/ksm/metric/metric_test.go
@@ -73,6 +73,25 @@ func TestGetDeploymentNameForReplicaSet_ValidName(t *testing.T) {
 	assert.Equal(t, expectedValue, fetchedValue)
 }
 
+func TestGetDeploymentNameForReplicaSet_ErrorOnEmptyData(t *testing.T) {
+	raw := definition.RawGroups{
+		"replicaset": {
+			"kube-state-metrics-4044341274": definition.RawMetrics{
+				"kube_replicaset_created": prometheus.Metric{
+					Value: prometheus.GaugeValue(1507117436),
+					Labels: map[string]string{
+						"namespace":  "kube-system",
+						"replicaset": "",
+					},
+				},
+			},
+		},
+	}
+	fetchedValue, err := GetDeploymentNameForReplicaSet()("replicaset", "kube-state-metrics-4044341274", raw)
+	assert.EqualError(t, err, "error generating deployment name for replica set. replicaset field is empty")
+	assert.Empty(t, fetchedValue)
+}
+
 func TestGetDeploymentNameForPod_CreatedByReplicaSet(t *testing.T) {
 	expectedValue := "fluentd-elasticsearch"
 	fetchedValue, err := GetDeploymentNameForPod()("pod", "fluentd-elasticsearch-jnqb7", rawGroups)
@@ -98,6 +117,40 @@ func TestGetDeploymentNameForPod_NotCreatedByReplicaSet(t *testing.T) {
 
 	fetchedValue, err := GetDeploymentNameForPod()("pod", rawEntityID, raw)
 	assert.Nil(t, err)
+	assert.Empty(t, fetchedValue)
+}
+
+func TestGetDeploymentNameForPod_ErrorOnEmptyData(t *testing.T) {
+	rawEntityID := "kube-addon-manager-minikube"
+	raw := definition.RawGroups{
+		"pod": {
+			"kube-addon-manager-minikube": definition.RawMetrics{
+				"kube_pod_info": prometheus.Metric{
+					Value: prometheus.GaugeValue(1507117436),
+					Labels: map[string]string{
+						"created_by_name": "newrelic-infra-monitoring",
+						"created_by_kind": "", // Empty created_by_kind
+					},
+				},
+			},
+		},
+	}
+
+	fetchedValue, err := GetDeploymentNameForPod()("pod", rawEntityID, raw)
+	assert.EqualError(t, err, "error generating deployment name for pod. created_by_kind field is empty")
+	assert.Empty(t, fetchedValue)
+
+	// Empty created_by_name
+	m := raw["pod"]["kube-addon-manager-minikube"]["kube_pod_info"].(prometheus.Metric)
+	m.Labels = map[string]string{
+		"created_by_name": "",
+		"created_by_kind": "DaemonSet",
+	}
+
+	raw["pod"]["kube-addon-manager-minikube"]["kube_pod_info"] = m
+
+	fetchedValue, err = GetDeploymentNameForPod()("pod", rawEntityID, raw)
+	assert.EqualError(t, err, "error generating deployment name for pod. created_by_name field is empty")
 	assert.Empty(t, fetchedValue)
 }
 
@@ -171,6 +224,65 @@ func TestGetDeploymentNameForContainer_NotCreatedByReplicaSet(t *testing.T) {
 	}
 	fetchedValue, err := GetDeploymentNameForContainer()("container", podRawID, raw)
 	assert.Nil(t, err)
+	assert.Empty(t, fetchedValue)
+}
+
+func TestGetDeploymentNameForContainer_ErrorOnMissingData(t *testing.T) {
+	podRawID := "kube-system_fluentd-elasticsearch-jnqb7_kube-state-metrics"
+	raw := definition.RawGroups{
+		"pod": {
+			"kube-system_kube-addon-manager-minikube": definition.RawMetrics{
+				"kube_pod_info": prometheus.Metric{
+					Value: prometheus.GaugeValue(1507117436),
+					Labels: map[string]string{
+						"namespace": "kube-system",
+						"node":      "minikube",
+						"pod":       "kube-addon-manager-minikube",
+					},
+				},
+			},
+		},
+		"container": {
+			podRawID: definition.RawMetrics{
+				"kube_pod_container_info": prometheus.Metric{
+					Value: prometheus.GaugeValue(1),
+					Labels: map[string]string{
+						"container": "kube-state-metrics",
+						"namespace": "kube-system",
+						"pod":       "kube-addon-manager-minikube",
+					},
+				},
+			},
+		},
+	}
+
+	// created_by_kind is empty
+	raw["pod"]["kube-system_kube-addon-manager-minikube"]["kube_pod_info"].(prometheus.Metric).Labels["created_by_name"] = "newrelic-infra-monitoring"
+
+	fetchedValue, err := GetDeploymentNameForContainer()("container", podRawID, raw)
+	assert.EqualError(t, err, "error generating deployment name for container. created_by_kind field is missing")
+	assert.Empty(t, fetchedValue)
+
+	// created_by_name is empty
+	raw["pod"]["kube-system_kube-addon-manager-minikube"]["kube_pod_info"].(prometheus.Metric).Labels["created_by_name"] = ""
+	raw["pod"]["kube-system_kube-addon-manager-minikube"]["kube_pod_info"].(prometheus.Metric).Labels["created_by_kind"] = "DaemonSet"
+
+	fetchedValue, err = GetDeploymentNameForContainer()("container", podRawID, raw)
+	assert.EqualError(t, err, "error generating deployment name for container. created_by_name field is missing")
+	assert.Empty(t, fetchedValue)
+
+	// Missing created_by_kind and created_by_name
+	m := raw["pod"]["kube-system_kube-addon-manager-minikube"]["kube_pod_info"].(prometheus.Metric)
+	m.Labels = map[string]string{
+		"namespace": "kube-system",
+		"node":      "minikube",
+		"pod":       "kube-addon-manager-minikube",
+	}
+
+	raw["pod"]["kube-system_kube-addon-manager-minikube"]["kube_pod_info"] = m
+
+	fetchedValue, err = GetDeploymentNameForContainer()("container", podRawID, raw)
+	assert.EqualError(t, err, "error generating deployment name for container. created_by_kind field is missing")
 	assert.Empty(t, fetchedValue)
 }